+++ src/bin/cargo-fm.rs
@@ -1,0 +1,6 @@
+#[path = "../main.rs"]
+mod cargo_feature_manifest;
+
+fn main() {
+    cargo_feature_manifest::cli_main();
+}