+++ tests/common/mod.rs
@@ -26,6 +26,13 @@ pub fn run_command(args: &[&str]) -> Output {
         .expect("failed to run cargo-feature-manifest")
 }
 
+pub fn run_short_command(args: &[&str]) -> Output {
+    Command::new(env!("CARGO_BIN_EXE_cargo-fm"))
+        .args(args)
+        .output()
+        .expect("failed to run cargo-fm")
+}
+
 pub fn normalize(text: &[u8]) -> String {
     String::from_utf8_lossy(text).replace("\r\n", "\n")
 }